// apps/admin/src/api/generated/schemas/emailLogResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { EmailLogResponseCreatedAt } from './emailLogResponseCreatedAt';
import type { EmailLogResponseUpdatedAt } from './emailLogResponseUpdatedAt';
import type { EmailLogResponseBodyPreview } from './emailLogResponseBodyPreview';
import type { EmailStatus } from './emailStatus';
import type { EmailLogResponseSentAt } from './emailLogResponseSentAt';
import type { EmailLogResponseErrorMessage } from './emailLogResponseErrorMessage';
import type { EmailLogResponseTemplateName } from './emailLogResponseTemplateName';
import type { EmailLogResponseScenario } from './emailLogResponseScenario';

/**
 * Schema for email log response.
 */
export interface EmailLogResponse {
  /** Creation timestamp */
  createdAt?: EmailLogResponseCreatedAt;
  /** Last update timestamp */
  updatedAt?: EmailLogResponseUpdatedAt;
  /** Unique identifier for the resource */
  id: string;
  tenantId: string;
  recipient: string;
  subject: string;
  bodyPreview?: EmailLogResponseBodyPreview;
  status: EmailStatus;
  sentAt?: EmailLogResponseSentAt;
  errorMessage?: EmailLogResponseErrorMessage;
  retryCount?: number;
  templateName?: EmailLogResponseTemplateName;
  scenario?: EmailLogResponseScenario;
}